requirements.txt
altair==5.4.1
annotated-types==0.7.0
attrs==24.2.0
beautifulsoup4==4.12.3
black==24.10.0
blinker==1.8.2
blis==1.0.1
cachetools==5.5.0
catalogue==2.0.10
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
cloudpathlib==0.20.0
confection==0.1.5
cymem==2.0.8
filelock==3.16.1
flake8==7.1.1
frozendict==2.4.6
fsspec==2024.10.0
gitdb==4.0.11
GitPython==3.1.43
html5lib==1.1
huggingface-hub==0.26.1
idna==3.10
iniconfig==2.0.0
isort==5.13.2
Jinja2==3.1.4
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
langcodes==3.4.1
language_data==1.2.0
lxml==5.3.0
marisa-trie==1.2.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mccabe==0.7.0
mdurl==0.1.2
mpmath==1.3.0
multitasking==0.0.11
murmurhash==1.0.10
mypy-extensions==1.0.0
narwhals==1.10.0
networkx==3.4.2
nltk==3.9.1
numpy==2.0.2
packaging==24.1
pandas==2.2.3
pathspec==0.12.1
peewee==3.17.7
pillow==10.4.0
platformdirs==4.3.6
pluggy==1.5.0
praw==7.8.1
prawcore==2.4.0
preshed==3.0.9
protobuf==5.28.3
pyarrow==17.0.0
pycodestyle==2.12.1
pydantic==2.9.2
pydantic_core==2.23.4
pydeck==0.9.1
pyflakes==3.2.0
Pygments==2.18.0
pytest==8.3.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
referencing==0.35.1
regex==2024.9.11
requests==2.32.3
rich==13.9.3
rpds-py==0.20.0
safetensors==0.4.5
scikit-learn==1.5.2
scipy==1.14.1
shellingham==1.5.4
six==1.16.0
smart-open==7.0.5
smmap==5.0.1
soupsieve==2.6
spacy==3.8.2
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.4.8
streamlit==1.39.0
sympy==1.13.1
tenacity==9.0.0
thinc==8.3.2
threadpoolctl==3.5.0
tokenizers==0.20.1
toml==0.10.2
torch==2.5.0
tornado==6.4.1
tqdm==4.66.5
transformers==4.46.0
typer==0.12.5
typing_extensions==4.12.2
tzdata==2024.2
update-checker==0.18.0
urllib3==2.2.3
wasabi==1.1.3
weasel==0.4.1
webencodings==0.5.1
websocket-client==1.8.0
wrapt==1.16.0
yfinance==0.2.48
